requirements.txt
numpy
flask
schedule
mysql-connector
python-dotenv
pillow
requests
beautifulsoup4